+++ include/dto/sqs/SQSQueueUpdateRequest.h
@@ -7,31 +7,41 @@
 
 struct SQSQueueUpdateRequest {
 
-public:
-
     QString queueArn;
 
-    long retentionPeriod;
+    QString deadLetterQueueArn;
 
-//    long maxMessageSize;
+    long retentionPeriod{};
 
-    long visibilityTimeout;
+    long maxMessageSize{};
 
-    long delay;
+    long visibilityTimeout{};
 
-    void FromJson(const QJsonDocument& jsonDoc) {
-        queueArn = jsonDoc["queueArn"].toString();
+    long delay{};
+
+    long maxRetries{};
+
+    QString owner;
+
+    void FromJson(const QJsonDocument &jsonDoc) {
+        queueArn = jsonDoc["arn"].toString();
+        deadLetterQueueArn = jsonDoc["deadLetterQueueArn"].toString();
+        maxRetries = jsonDoc["maxRetries"].toInt();
         retentionPeriod = jsonDoc["retentionPeriod"].toInteger();
         visibilityTimeout = jsonDoc["visibilityTimeout"].toInteger();
         delay = jsonDoc["delay"].toInteger();
+        owner = jsonDoc["owner"].toString();
     }
 
     [[nodiscard]] QByteArray ToJson() const {
         QJsonObject jObject;
-        jObject["queueArn"] = queueArn;
+        jObject["arn"] = queueArn;
+        jObject["deadLetterQueueArn"] = deadLetterQueueArn;
         jObject["retentionPeriod"] = static_cast<qint64>(retentionPeriod);
         jObject["visibilityTimeout"] = static_cast<qint64>(visibilityTimeout);
         jObject["delay"] = static_cast<qint64>(delay);
+        jObject["maxRetries"] = static_cast<qint64>(maxRetries);
+        jObject["owner"] = owner;
         const QJsonDocument jDoc(jObject);
         return jDoc.toJson();
     }