+++ src/modules/s3/S3BucketList.cpp
@@ -77,6 +77,13 @@ S3BucketList::S3BucketList(const QString &title, QWidget *parent) : BasePage(par
     // Connect paging changes
     connect(_tableView, &PageableTable::ReloadTable, this, &S3BucketList::LoadContent);
 
+    // Add details shortcut
+    connect(_tableView, &PageableTable::ShowDetailsSignal, this, [this](const QModelIndex &index) {
+        const auto bucketName = _tableView->GetValue<QString>(index, 0);
+        S3BucketEditDialog dialog(bucketName);
+        dialog.exec();
+    });
+
     // Set up the layout for the individual content pages
     const auto layout = new QVBoxLayout(this);
     layout->addLayout(toolBar, 0);