+++ src/modules/sqs/SQSQueueList.cpp
@@ -24,8 +24,7 @@ SQSQueueList::SQSQueueList(const QString &title, QWidget *parent) : BasePage(par
     addButton->setToolTip("Add a new queue");
     connect(addButton, &QPushButton::clicked, [this]() {
         bool ok;
-        if (const QString text = QInputDialog::getText(nullptr, "Queue Name", "Queue name:", QLineEdit::Normal, "", &ok)
-            ; ok && !text.isEmpty()) {
+        if (const QString text = QInputDialog::getText(nullptr, "Queue Name", "Queue name:", QLineEdit::Normal, "", &ok); ok && !text.isEmpty()) {
             _sqsService->AddQueue(text);
         }
     });
@@ -52,7 +51,6 @@ SQSQueueList::SQSQueueList(const QString &title, QWidget *parent) : BasePage(par
         LoadContent();
     });
 
-    //toolBar->addWidget(backButton);
     toolBar->addWidget(titleLabel);
     toolBar->addWidget(spacer);
     toolBar->addWidget(addButton);
@@ -87,6 +85,13 @@ SQSQueueList::SQSQueueList(const QString &title, QWidget *parent) : BasePage(par
 
     // Add context menu
     connect(_tableView, &PageableTable::ContextMenuRequested, this, &SQSQueueList::ShowContextMenu);
+
+    // Add details shortcut
+    connect(_tableView, &PageableTable::ShowDetailsSignal, this, [this](const QModelIndex &index) {
+        const auto queueArn = _tableView->GetValue<QString>(index, 8);
+        SQSQueueDetailsDialog dialog(queueArn);
+        dialog.exec();
+    });
 
     // Set up the layout for the individual content pages
     const auto layout = new QVBoxLayout(this);
@@ -128,7 +133,6 @@ void SQSQueueList::ShowContextMenu(const QPoint &pos) {
 
     const QModelIndex index = _tableView->GetIndexFromPosition(pos);
     const long total = _tableView->GetValue<long>(index, 1) + _tableView->GetValue<long>(index, 2) + _tableView->GetValue<long>(index, 3);
-    const bool isDlq = _tableView->GetValue<bool>(index, 9);
 
     auto *menu = new ContextMenu(this);
 
@@ -145,8 +149,8 @@ void SQSQueueList::ShowContextMenu(const QPoint &pos) {
     purgeAction->setEnabled(total > 0);
 
     QAction *redriveAction = menu->addAction(IconUtils::GetIcon("redrive"), "Redrive Queue");
-    redriveAction->setToolTip("Redrive all messages");
-    redriveAction->setEnabled(isDlq && total > 0);
+    redriveAction->setToolTip("Redrive all messages. If the selected queue is a DLQ, it will perform an redrive, otherwise it will set all messages to status 'INITIAL'.");
+    redriveAction->setEnabled(total > 0);
 
     menu->addSeparator();